refactor: simplify home feature slice.
