+++ lib/features/home/home.dart
@@ -1,2 +1,2 @@
-export 'presentation/pages/home_page.dart';
-export 'presentation/providers/navigation_visible_provider.dart';
+export 'views/home_page.dart';
+export 'providers/navigation_visible_provider.dart';
