--- lib/features/home/presentation/presentation.dart
@@ -1,8 +1,0 @@
-// Models
-export 'models/destinations.dart';
-
-// Pages.
-export 'pages/home_page.dart';
-
-// Providers
-export 'providers/navigation_visible_provider.dart';
