+++ lib/features/home/providers/navigation_visible_provider.dart
